refactor(components): move form
move form rendering in separate component
